Task 11

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
 */
 int main(void)
 {
-	unsigned prime 612852475143, div;
+	long prime 612852475143, div;
 
 	while (div < (prime / 2))
 	{
